Refactored configuration tests
